+++ src/compas_timber/connections/analyzers.py
@@ -10,6 +10,7 @@ from compas.tolerance import TOL
 import compas_timber.connections  # noqa: F401
 import compas_timber.elements  # noqa: F401
 from compas_timber.connections import GenericJoint
+from compas_timber.connections import JointTopology
 
 
 class Cluster(object):
@@ -53,6 +54,26 @@ class Cluster(object):
     def location(self):
         # type: () -> compas.geometry.Point
         return self.joints[0].location
+
+    @property
+    def topology(self):
+        """Returns the topology of the joint if there is only one joint, otherwise TOPO_UNKNOWN."""
+        # TODO: will we ever have clusters from non-GenericJoints? if so then we coult have a joint in a cluster with TOPO_Y or TOPO_K
+        # TOPO_Y + TOPO_I = TOPO_Y
+        # TOPO_Y + TOPO_L = TOPO_Y
+        # TOPO_Y + TOPO_T = TOPO_K
+        # TOPO_K + TOPO_I = TOPO_K
+        # TOPO_K + TOPO_L = TOPO_K ...
+        if len(self.joints) == 0:
+            return JointTopology.TOPO_UNKNOWN
+        if len(self.joints) == 1:
+            return self.joints[0].topology
+        if any([j.topology not in [JointTopology.TOPO_L, JointTopology.TOPO_I, JointTopology.TOPO_T, JointTopology.TOPO_X] for j in self.joints]):
+            return JointTopology.TOPO_UNKNOWN
+        if any([j.topology == JointTopology.TOPO_T or j.topology == JointTopology.TOPO_X for j in self.joints]):
+            return JointTopology.TOPO_K
+        return JointTopology.TOPO_Y
+
 
 
 class BeamGroupAnalyzer(object):