+++ src/compas_timber/connections/plate_joint.py
@@ -258,7 +258,7 @@ class PlateJoint(Joint):
 
         self.a_planes = self.plate_a.planes
         self.a_outlines = self.plate_a.outlines
-        if self.topology == JointTopology.TOPO_L:
+        if self.topology == JointTopology.TOPO_EDGE_EDGE:
             if dot_vectors(self.plate_a.frame.normal, get_polyline_segment_perpendicular_vector(self.plate_b.outline_a, self.b_segment_index)) < 0:
                 self.a_planes = self.plate_a.planes[::-1]
                 self.a_outlines = self.plate_a.outlines[::-1]