+++ tests/compas_timber/test_plate_joint.py
@@ -16,7 +16,7 @@ def test_plate_L_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_a, plate_b)
-    assert topo_results[0] == JointTopology.TOPO_L, "Expected L-joint topology"
+    assert topo_results[0] == JointTopology.TOPO_EDGE_EDGE, "Expected L-joint topology"
     assert topo_results[1][0] == plate_a, "Expected plate_a as first plate in topology result"
     assert topo_results[2][0] == plate_b, "Expected plate_b as second plate in topology result"
     assert topo_results[1][1] == 1, "Expected connection segment at index = 1"
@@ -33,7 +33,7 @@ def test_plate_T_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_a, plate_b)
-    assert topo_results[0] == JointTopology.TOPO_T, "Expected T-joint topology"
+    assert topo_results[0] == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
     assert topo_results[1][0] == plate_b, "Expected plate_a as first plate in topology result"
     assert topo_results[2][0] == plate_a, "Expected plate_b as second plate in topology result"
     assert topo_results[1][1] == 0, "Expected connection segment at index = 1"
@@ -50,7 +50,7 @@ def test_reversed_plate_T_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_b, plate_a)
-    assert topo_results[0] == JointTopology.TOPO_T, "Expected T-joint topology"
+    assert topo_results[0] == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
     assert topo_results[1][0] == plate_b, "Expected plate_a as first plate in topology result"
     assert topo_results[2][0] == plate_a, "Expected plate_b as second plate in topology result"
     assert topo_results[1][1] == 0, "Expected connection segment at index = 1"
@@ -76,7 +76,7 @@ def test_three_plate_topos():
     topo_results.append(cs.find_topology(plate_a, plate_c))
 
     assert len(topo_results) == 3, "Expected three topology results"
-    assert all(tr[0] == JointTopology.TOPO_L for tr in topo_results), "Expected all topology results to be L-joints"
+    assert all(tr[0] == JointTopology.TOPO_EDGE_EDGE for tr in topo_results), "Expected all topology results to be L-joints"
 
 
 def test_three_plate_mix_topos():
@@ -98,9 +98,9 @@ def test_three_plate_mix_topos():
     topo_results.append(cs.find_topology(plate_a, plate_c))
 
     assert len(topo_results) == 3, "Expected three topology results"
-    assert topo_results[0][0] == JointTopology.TOPO_T, "Expected first topology result to be T-joint"
-    assert topo_results[1][0] == JointTopology.TOPO_L, "Expected second topology result to be L-joint"
-    assert topo_results[2][0] == JointTopology.TOPO_L, "Expected third topology result to be L-joint"
+    assert topo_results[0][0] == JointTopology.TOPO_EDGE_FACE, "Expected first topology result to be T-joint"
+    assert topo_results[1][0] == JointTopology.TOPO_EDGE_EDGE, "Expected second topology result to be L-joint"
+    assert topo_results[2][0] == JointTopology.TOPO_EDGE_EDGE, "Expected third topology result to be L-joint"
 
 
 def test_simple_joint_and_reset():
@@ -157,9 +157,9 @@ def test_three_plate_joints():
         kwargs = {"topology": tr[0], "a_segment_index": tr[1][1], "b_segment_index": tr[2][1]}
         if tr[0] == JointTopology.TOPO_UNKNOWN:
             continue
-        elif tr[0] == JointTopology.TOPO_L:
+        elif tr[0] == JointTopology.TOPO_EDGE_EDGE:
             joints.append(PlateMiterJoint(tr[1][0], tr[2][0], **kwargs))
-        elif tr[0] == JointTopology.TOPO_T:
+        elif tr[0] == JointTopology.TOPO_EDGE_FACE:
             joints.append(PlateButtJoint(tr[1][0], tr[2][0], **kwargs))
 
     assert len(joints) == 3, "Expected three joints"
@@ -191,9 +191,9 @@ def test_three_plate_joints_mix_topo():
 
         if tr[0] == JointTopology.TOPO_UNKNOWN:
             continue
-        elif tr[0] == JointTopology.TOPO_L:
+        elif tr[0] == JointTopology.TOPO_EDGE_EDGE:
             joints.append(PlateMiterJoint(tr[1][0], tr[2][0], **kwargs))
-        elif tr[0] == JointTopology.TOPO_T:
+        elif tr[0] == JointTopology.TOPO_EDGE_FACE:
             joints.append(PlateButtJoint(tr[1][0], tr[2][0], **kwargs))
 
     assert len(joints) == 3, "Expected three joints"